Rename module to 'models'
