+++ homeassistant/components/spotify/browse_media.py
@@ -19,8 +19,8 @@ from homeassistant.components.media_player import (
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.config_entry_oauth2_flow import OAuth2Session
 
-from .config_entry import HomeAssistantSpotifyData
 from .const import DOMAIN, MEDIA_PLAYER_PREFIX, MEDIA_TYPE_SHOW, PLAYABLE_MEDIA_TYPES
+from .models import HomeAssistantSpotifyData
 from .util import fetch_image_url
 
 BROWSE_LIMIT = 48
