+++ homeassistant/components/spotify/models.py
